server/requirements.txt
Django==3.2.11
PyMySQL==1.0.2
djangorestframework==3.13.0
django-cors-headers==3.13.0
Pillow==9.1.1
psutil==5.9.4
openai